Move context menu actions from `CommitListItem` to `CommitList`
