+++ app/src/ui/drag-elements/commit-drag-element.tsx
@@ -178,11 +178,6 @@ export class CommitDragElement extends React.Component<
             commit={commit}
             selectedCommits={selectedCommits}
             emoji={emoji}
-            canBeUndone={false}
-            canBeAmended={false}
-            canBeResetTo={false}
-            canBeCheckedOut={false}
-            isLocal={false}
             showUnpushedIndicator={false}
           />
         </div>
